backend/requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
click==8.1.3
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
Django==4.1.7
django-cors-headers==3.13.0
djangorestframework==3.14.0
dnspython==2.3.0
executing==1.2.0
fonttools==4.39.3
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.21.2
ipython==8.10.0
jedi==0.18.2
joblib==1.2.0
jupyter_client==8.0.3
jupyter_core==5.2.0
kiwisolver==1.4.4
matplotlib==3.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.24.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.0.0
prompt-toolkit==3.0.37
psutil==5.9.4
pure-eval==0.2.2
Pygments==2.14.0
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
pywin32==305
pyzmq==25.0.0
regex==2022.10.31
scipy==1.10.1
six==1.16.0
sklearn==0.0.post1
sqlparse==0.4.3
stack-data==0.6.2
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
tzdata==2022.7
wcwidth==0.2.6
zipp==3.14.0
